New Resource: azurerm_cognitive_account_connection_aad, azurerm_cognitive_account_connection_api_key, azurerm_cognitive_account_connection_custom_keys, azurerm_cognitive_account_connection_oauth2
