--- examples/cognitive/account-connection/api-key/README.md
@@ -1,10 +1,0 @@
-# Cognitive Account Connection - API Key Authentication
-
-This example provisions a Cognitive Services Account Connection using API Key authentication to connect an AI Services account to an OpenAI account.
-
-## Resources Provisioned
-
-* Resource Group
-* Cognitive Services Account (AI Services)
-* Cognitive Services Account (OpenAI)
-* Cognitive Account Connection with API Key authentication
