--- examples/cognitive/account-connection/custom-keys/README.md
@@ -1,10 +1,0 @@
-# Cognitive Account Connection - Custom Keys Authentication
-
-This example provisions a Cognitive Services Account Connection using Custom Keys authentication to connect an AI Services account to an OpenAI account.
-
-## Resources Provisioned
-
-* Resource Group
-* Cognitive Services Account (AI Services)
-* Cognitive Services Account (OpenAI)
-* Cognitive Account Connection with Custom Keys authentication
