--- examples/cognitive/account-connection/oauth2/README.md
@@ -1,11 +1,0 @@
-# Cognitive Account Connection - OAuth2 Authentication
-
-This example provisions a Cognitive Services Account Connection using OAuth2 authentication to connect an AI Services account to an Azure Storage Blob.
-
-## Resources Provisioned
-
-* Resource Group
-* Cognitive Services Account (AI Services)
-* Storage Account
-* Storage Container
-* Cognitive Account Connection with OAuth2 authentication
