+++ internal/services/cognitive/cognitive_account_connection_api_key_resource_test.go
@@ -1,0 +1,197 @@
+package cognitive_test
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/hashicorp/go-azure-helpers/lang/pointer"
+	"github.com/hashicorp/go-azure-sdk/resource-manager/cognitive/2025-06-01/accountconnectionresource"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/acceptance"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/acceptance/check"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/clients"
+	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
+)
+
+type CognitiveAccountConnectionApiKeyResource struct{}
+
+func TestAccCognitiveAccountConnectionApiKey_basic(t *testing.T) {
+	data := acceptance.BuildTestData(t, "azurerm_cognitive_account_connection_api_key", "test")
+	r := CognitiveAccountConnectionApiKeyResource{}
+
+	data.ResourceTest(t, r, []acceptance.TestStep{
+		{
+			Config: r.basic(data),
+			Check: acceptance.ComposeTestCheckFunc(
+				check.That(data.ResourceName).ExistsInAzure(r),
+			),
+		},
+		data.ImportStep("api_key", "metadata"),
+	})
+}
+
+func TestAccCognitiveAccountConnectionApiKey_requiresImport(t *testing.T) {
+	data := acceptance.BuildTestData(t, "azurerm_cognitive_account_connection_api_key", "test")
+	r := CognitiveAccountConnectionApiKeyResource{}
+
+	data.ResourceTest(t, r, []acceptance.TestStep{
+		{
+			Config: r.basic(data),
+			Check: acceptance.ComposeTestCheckFunc(
+				check.That(data.ResourceName).ExistsInAzure(r),
+			),
+		},
+		data.RequiresImportErrorStep(r.requiresImport),
+	})
+}
+
+func TestAccCognitiveAccountConnectionApiKey_update(t *testing.T) {
+	data := acceptance.BuildTestData(t, "azurerm_cognitive_account_connection_api_key", "test")
+	r := CognitiveAccountConnectionApiKeyResource{}
+
+	data.ResourceTest(t, r, []acceptance.TestStep{
+		{
+			Config: r.basic(data),
+			Check: acceptance.ComposeTestCheckFunc(
+				check.That(data.ResourceName).ExistsInAzure(r),
+			),
+		},
+		data.ImportStep("api_key", "metadata"),
+		{
+			Config: r.updated(data),
+			Check: acceptance.ComposeTestCheckFunc(
+				check.That(data.ResourceName).ExistsInAzure(r),
+			),
+		},
+		data.ImportStep("api_key", "metadata"),
+	})
+}
+
+func (r CognitiveAccountConnectionApiKeyResource) Exists(ctx context.Context, clients *clients.Client, state *pluginsdk.InstanceState) (*bool, error) {
+	id, err := accountconnectionresource.ParseConnectionID(state.ID)
+	if err != nil {
+		return nil, err
+	}
+
+	resp, err := clients.Cognitive.AccountConnectionResourceClient.AccountConnectionsGet(ctx, *id)
+	if err != nil {
+		return nil, fmt.Errorf("retrieving %s: %+v", *id, err)
+	}
+
+	return pointer.To(resp.Model != nil), nil
+}
+
+func (r CognitiveAccountConnectionApiKeyResource) template(data acceptance.TestData) string {
+	return fmt.Sprintf(`
+resource "azurerm_resource_group" "test" {
+  name     = "acctestRG-cognitive-ac-%[1]d"
+  location = "%[2]s"
+}
+
+resource "azurerm_cognitive_account" "test" {
+  name                       = "acctest-aiservices-%[1]d"
+  location                   = azurerm_resource_group.test.location
+  resource_group_name        = azurerm_resource_group.test.name
+  kind                       = "AIServices"
+  sku_name                   = "S0"
+  project_management_enabled = true
+  custom_subdomain_name      = "acctestaiservices-%[1]d"
+
+  identity {
+    type = "SystemAssigned"
+  }
+}
+
+resource "azurerm_cognitive_account" "openai" {
+  name                = "acctest-openai-%[1]d"
+  location            = azurerm_resource_group.test.location
+  resource_group_name = azurerm_resource_group.test.name
+  kind                = "OpenAI"
+  sku_name            = "S0"
+
+  identity {
+    type = "SystemAssigned"
+  }
+}
+`, data.RandomInteger, data.Locations.Primary)
+}
+
+func (r CognitiveAccountConnectionApiKeyResource) basic(data acceptance.TestData) string {
+	return fmt.Sprintf(`
+provider "azurerm" {
+  features {}
+}
+
+%[1]s
+
+resource "azurerm_cognitive_account_connection_api_key" "test" {
+  name                 = "acctest-conn-%[2]d"
+  cognitive_account_id = azurerm_cognitive_account.test.id
+  category             = "AzureOpenAI"
+  target               = azurerm_cognitive_account.openai.endpoint
+  api_key              = azurerm_cognitive_account.openai.primary_access_key
+
+  metadata = {
+    apiType    = "Azure"
+    resourceId = azurerm_cognitive_account.openai.id
+    location   = azurerm_cognitive_account.openai.location
+  }
+}
+`, r.template(data), data.RandomInteger)
+}
+
+func (r CognitiveAccountConnectionApiKeyResource) requiresImport(data acceptance.TestData) string {
+	return fmt.Sprintf(`
+%[1]s
+
+resource "azurerm_cognitive_account_connection_api_key" "import" {
+  name                 = azurerm_cognitive_account_connection_api_key.test.name
+  cognitive_account_id = azurerm_cognitive_account_connection_api_key.test.cognitive_account_id
+  category             = azurerm_cognitive_account_connection_api_key.test.category
+  target               = azurerm_cognitive_account_connection_api_key.test.target
+  api_key              = azurerm_cognitive_account_connection_api_key.test.api_key
+
+  metadata = {
+    apiType    = azurerm_cognitive_account_connection_api_key.test.metadata.apiType
+    resourceId = azurerm_cognitive_account_connection_api_key.test.metadata.resourceId
+    location   = azurerm_cognitive_account_connection_api_key.test.metadata.location
+  }
+}
+`, r.basic(data))
+}
+
+func (r CognitiveAccountConnectionApiKeyResource) updated(data acceptance.TestData) string {
+	return fmt.Sprintf(`
+provider "azurerm" {
+  features {}
+}
+
+%[1]s
+
+resource "azurerm_cognitive_account" "openai2" {
+  name                = "acctest-openai2-%[2]d"
+  location            = azurerm_resource_group.test.location
+  resource_group_name = azurerm_resource_group.test.name
+  kind                = "OpenAI"
+  sku_name            = "S0"
+
+  identity {
+    type = "SystemAssigned"
+  }
+}
+
+resource "azurerm_cognitive_account_connection_api_key" "test" {
+  name                 = "acctest-conn-%[2]d"
+  cognitive_account_id = azurerm_cognitive_account.test.id
+  category             = "AzureOpenAI"
+  target               = azurerm_cognitive_account.openai2.endpoint
+  api_key              = azurerm_cognitive_account.openai2.primary_access_key
+
+  metadata = {
+    apiType    = "Azure"
+    resourceId = azurerm_cognitive_account.openai2.id
+    location   = azurerm_cognitive_account.openai2.location
+  }
+}
+`, r.template(data), data.RandomInteger)
+}
