+++ internal/services/cognitive/registration.go
@@ -60,7 +60,10 @@ func (r Registration) DataSources() []sdk.DataSource {
 // Resources returns a list of Resources supported by this Service
 func (r Registration) Resources() []sdk.Resource {
 	resources := []sdk.Resource{
-		CognitiveAccountConnectionResource{},
+		CognitiveAccountConnectionAADResource{},
+		CognitiveAccountConnectionApiKeyResource{},
+		CognitiveAccountConnectionCustomKeysResource{},
+		CognitiveAccountConnectionOAuth2Resource{},
 		CognitiveAccountProjectResource{},
 		CognitiveAccountRaiPolicyResource{},
 		CognitiveDeploymentResource{},
