+++ website/docs/r/cognitive_account_connection_oauth2.html.markdown
@@ -1,18 +1,16 @@
 ---
 subcategory: "Cognitive Services"
 layout: "azurerm"
-page_title: "Azure Resource Manager: azurerm_cognitive_account_connection"
+page_title: "Azure Resource Manager: azurerm_cognitive_account_connection_oauth2"
 description: |-
-  Manages a Cognitive Services Account Connection.
+  Manages a Cognitive Services Account Connection with OAuth2 authentication.
 ---
 
-# azurerm_cognitive_account_connection
+# azurerm_cognitive_account_connection_oauth2
 
-Manages a Cognitive Services Account Connection.
+Manages a Cognitive Services Account Connection with OAuth2 authentication.
 
 ## Example Usage
-
-This example provisions a Cognitive Services Account Connection with AAD authentication. Additional examples of how to use the `azurerm_cognitive_account_connection` resource can be found in [the `./examples/cognitive/account-connection` directory within the GitHub Repository](https://github.com/hashicorp/terraform-provider-azurerm/tree/main/examples/cognitive/account-connection).
 
 ```hcl
 resource "azurerm_resource_group" "example" {
@@ -48,16 +46,22 @@ resource "azurerm_storage_container" "example" {
   container_access_type = "private"
 }
 
-resource "azurerm_cognitive_account_connection" "example" {
+resource "azurerm_cognitive_account_connection_oauth2" "example" {
   name                 = "example-connection"
   cognitive_account_id = azurerm_cognitive_account.example.id
-  authentication_type  = "AAD"
   category             = "AzureBlob"
   target               = azurerm_storage_account.example.primary_blob_endpoint
 
   metadata = {
-    accountName   = azurerm_storage_account.example.name
     containerName = azurerm_storage_container.example.name
+    accountName   = azurerm_storage_account.example.name
+  }
+
+  oauth2 {
+    authentication_url = "https://login.microsoftonline.com/00000000-0000-0000-0000-000000000000/oauth2/v2.0/token"
+    client_id          = "00000000-0000-0000-0000-000000000000"
+    client_secret      = "example-client-secret"
+    tenant_id          = "00000000-0000-0000-0000-000000000000"
   }
 }
 ```
@@ -70,25 +74,13 @@ The following arguments are supported:
 
 * `cognitive_account_id` - (Required) The ID of the Cognitive Services Account. Changing this forces a new resource to be created.
 
-* `authentication_type` - (Required) The authentication type for the connection. Possible values are `AAD`, `ApiKey`, `CustomKeys`, and `OAuth2`.
-
-* `category` - (Required) The category of the connection. Possible values are `ADLSGen2`, `AIServices`, `AmazonMws`, `AmazonRdsForOracle`, `AmazonRdsForSqlServer`, `AmazonRedshift`, `AmazonS3Compatible`, `ApiKey`, `AzureBlob`, `AzureDataExplorer`, `AzureDatabricksDeltaLake`, `AzureMariaDb`, `AzureMySqlDb`, `AzureOneLake`, `AzureOpenAI`, `AzurePostgresDb`, `AzureSqlDb`, `AzureSqlMi`, `AzureSynapseAnalytics`, `AzureTableStorage`, `BingLLMSearch`, `Cassandra`, `CognitiveSearch`, `CognitiveService`, `Concur`, `ContainerRegistry`, `CosmosDb`, `CosmosDbMongoDbApi`, `Couchbase`, `CustomKeys`, `Db2`, `Drill`, `Dynamics`, `DynamicsAx`, `DynamicsCrm`, `Elasticsearch`, `Eloqua`, `FileServer`, `FtpServer`, `GenericContainerRegistry`, `GenericHttp`, `GenericRest`, `Git`, `GoogleAdWords`, `GoogleBigQuery`, `GoogleCloudStorage`, `Greenplum`, `Hbase`, `Hdfs`, `Hive`, `Hubspot`, `Impala`, `Informix`, `Jira`, `Magento`, `ManagedOnlineEndpoint`, `MariaDb`, `Marketo`, `MicrosoftAccess`, `MongoDbAtlas`, `MongoDbV2`, `MySql`, `Netezza`, `ODataRest`, `Odbc`, `Office365`, `OpenAI`, `Oracle`, `OracleCloudStorage`, `OracleServiceCloud`, `PayPal`, `Phoenix`, `Pinecone`, `PostgreSql`, `Presto`, `PythonFeed`, `QuickBooks`, `Redis`, `Responsys`, `S3`, `Salesforce`, `SalesforceMarketingCloud`, `SalesforceServiceCloud`, `SapBw`, `SapCloudForCustomer`, `SapEcc`, `SapHana`, `SapOpenHub`, `SapTable`, `Serp`, `Serverless`, `ServiceNow`, `Sftp`, `SharePointOnlineList`, `Shopify`, `Snowflake`, `Spark`, `SqlServer`, `Square`, `Sybase`, `Teradata`, `Vertica`, `WebTable`, `Xero`, and `Zoho`. Changing this forces a new resource to be created.
+* `category` - (Required) The category of the connection. Possible values include `AzureBlob`, `AzureOpenAI`, and other supported connection categories. Changing this forces a new resource to be created.
 
 * `metadata` - (Required) A mapping of metadata key-value pairs for the connection.
 
 * `target` - (Required) The target endpoint or resource for the connection.
 
-* `api_key` - (Optional) The API key for authentication. This field is sensitive.
-
-~> **Note:** `api_key` is required when `authentication_type` is set to `ApiKey`. `api_key` cannot be set together with `oauth2` or `custom_keys`.
-
-* `custom_keys` - (Optional) A mapping of custom keys for authentication. All values in this map are sensitive.
-
-~> **Note:** `custom_keys` is required when `authentication_type` is set to `CustomKeys`. `custom_keys` cannot be set together with `api_key` or `oauth2`.
-
-* `oauth2` - (Optional) An `oauth2` block as defined below.
-
-~> **Note:** `oauth2` is required when `authentication_type` is set to `OAuth2`. `oauth2` cannot be set together with `api_key` or `custom_keys`.
+* `oauth2` - (Required) An `oauth2` block as defined below.
 
 ---
 
@@ -130,7 +122,7 @@ The `timeouts` block allows you to specify [timeouts](https://developer.hashicor
 Cognitive Services Account Connections can be imported using the `resource id`, e.g.
 
 ```shell
-terraform import azurerm_cognitive_account_connection.example /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/group1/providers/Microsoft.CognitiveServices/accounts/account1/connections/connection1
+terraform import azurerm_cognitive_account_connection_oauth2.example /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/group1/providers/Microsoft.CognitiveServices/accounts/account1/connections/connection1
 ```
 
 ## API Providers
